feat(misc): dont generate defaultProject for non standalone workspaces
