+++ packages/angular/src/generators/application/angular-v14/application.ts
@@ -20,7 +20,6 @@ import {
   enableStrictTypeChecking,
   normalizeOptions,
   setApplicationStrictDefault,
-  setDefaultProject,
   updateAppComponentTemplate,
   updateComponentSpec,
   updateConfigFiles,
@@ -108,10 +107,6 @@ export async function applicationGenerator(
   await addUnitTestRunner(host, options);
   await addE2e(host, options);
   updateEditorTsConfig(host, options);
-
-  if (!options.skipDefaultProject) {
-    setDefaultProject(host, options);
-  }
 
   if (options.backendProject) {
     addProxyConfig(host, options);
