+++ packages/angular/src/generators/application/lib/index.ts
@@ -11,7 +11,6 @@ export * from './nrwl-home-tpl';
 export * from './remove-scaffolded-e2e';
 export * from './root-router-config';
 export * from './set-app-strict-default';
-export * from './set-default-project';
 export * from './update-component-spec';
 export * from './update-app-component-template';
 export * from './update-nx-component-template';
