+++ packages/angular/src/generators/application/schema.d.ts
@@ -24,7 +24,6 @@ export interface Schema {
   port?: number;
   setParserOptionsProject?: boolean;
   skipPackageJson?: boolean;
-  skipDefaultProject?: boolean;
   standalone?: boolean;
   rootProject?: boolean;
 }
