+++ packages/angular/src/generators/application/schema.json
@@ -142,11 +142,6 @@
       "description": "Whether to configure Tailwind CSS for the application.",
       "default": false
     },
-    "skipDefaultProject": {
-      "description": "Skip setting the project as the default project. When `false` (the default), the project is set as the default project only if there is no default project already set.",
-      "type": "boolean",
-      "default": false
-    },
     "standalone": {
       "description": "Generate an application that is setup to use standalone components.",
       "type": "boolean",
