+++ packages/angular/src/generators/remote/remote.ts
@@ -21,7 +21,6 @@ export async function remote(tree: Tree, options: Schema) {
   const appInstallTask = await applicationGenerator(tree, {
     ...options,
     routing: true,
-    skipDefaultProject: true,
     port,
   });
