+++ packages/expo/src/generators/application/lib/add-project.ts
@@ -24,14 +24,6 @@ export function addProject(host: Tree, options: NormalizedSchema) {
     projectConfiguration,
     options.standaloneConfig
   );
-
-  const workspace = readWorkspaceConfiguration(host);
-
-  if (!workspace.defaultProject) {
-    workspace.defaultProject = options.projectName;
-
-    updateWorkspaceConfiguration(host, workspace);
-  }
 }
 
 function getTargets(options: NormalizedSchema) {
