+++ packages/next/src/generators/application/lib/set-defaults.ts
@@ -9,10 +9,6 @@ import { NormalizedSchema } from './normalize-options';
 export function setDefaults(host: Tree, options: NormalizedSchema) {
   const workspace = readWorkspaceConfiguration(host);
 
-  if (!workspace.defaultProject) {
-    workspace.defaultProject = options.projectName;
-  }
-
   workspace.generators = workspace.generators || {};
   workspace.generators['@nrwl/next'] = workspace.generators['@nrwl/next'] || {};
   const prev = workspace.generators['@nrwl/next'];
