+++ packages/web/src/generators/application/application.spec.ts
@@ -26,13 +26,10 @@ describe('app', () => {
         standaloneConfig: false,
       });
       const workspaceJson = readJson(tree, '/workspace.json');
-      const nxJson = readJson<NxJsonConfiguration>(tree, 'nx.json');
-
       expect(workspaceJson.projects['my-app'].root).toEqual('apps/my-app');
       expect(workspaceJson.projects['my-app-e2e'].root).toEqual(
         'apps/my-app-e2e'
       );
-      expect(nxJson.defaultProject).toEqual('my-app');
     });
 
     it('should update tags and implicit dependencies', async () => {
